// src/app/shared/models/hint-model.ts
import { HintData } from 'desy-angular';

export class HintModel implements HintData{
  ariaControls: string;
  ariaCurrent: string;
  ariaDescribedBy: string;
  ariaDisabled: string;
  ariaErrorMessage: string;
  ariaExpanded: string;
  ariaHasPopup: string;
  ariaHidden: string;
  ariaLabel: string;
  ariaLabelledBy: string;
  ariaLive: string;
  classes: string;
  html: string;
  id: string;
  role: string;
  text: string;
  tabindex: string;
}